fix warnings/errors

+++ problems/Problem_014.hpp
@@ -18,8 +18,8 @@ inline double monteCarloPi( int n )
 
   for( int i = 0; i < n; ++i )
   {
-    double x = ( (float) rand() ) / (float) RAND_MAX;
-    double y = ( (float) rand() ) / (float) RAND_MAX;
+    double x = ( (double) rand() ) / (double) RAND_MAX;
+    double y = ( (double) rand() ) / (double) RAND_MAX;
 
     total++;
